+++ GameServerLib/Logic/Maps/TwistedTreeline.cs
@@ -243,7 +243,7 @@ namespace LeagueSandbox.GameServer.Logic.Maps
             return _turretItems[type];
         }
 
-        public override void Update(long diff)
+        public override void Update(float diff)
         {
             base.Update(diff);
